updated docstrings, type hints, and deleted duplicated fragments
