+++ ptls/data_load/iterable_processing_dataset.py
@@ -5,15 +5,17 @@ from ptls.data_load.feature_dict import FeatureDict
 class IterableProcessingDataset(FeatureDict, IterableDataset):
     def __init__(self):
         super().__init__()
-
         self._src = None
+        self._sequence_col = None
 
     def __call__(self, src):
         self._src = src
         return iter(self)
 
     def __iter__(self):
-        """For record transformation. Redefine __iter__ for filter"""
+        """
+        For record transformation. Redefine __iter__ for filter
+        """
         for rec in self._src:
             if type(rec) is tuple:
                 features = rec[0]
@@ -26,3 +28,11 @@ class IterableProcessingDataset(FeatureDict, IterableDataset):
 
     def process(self, features):
         raise NotImplementedError()
+
+    def get_sequence_col(self, rec):
+        if self._sequence_col is None:
+            arrays = [k for k, v in rec.items() if self.is_seq_feature(k, v)]
+            if len(arrays) == 0:
+                raise ValueError(f'Can not find field with sequence from record: {rec}')
+            self._sequence_col = arrays[0]
+        return self._sequence_col
